Padded input area and expanded it to fit input
Styled calculator input area to make it have a minimum height of 50px, but also made the numbers inside it wrap when they exceed the input area width. This makes the input area height dynamic based on how big the numbers inside are. Additionally added styling to the calculator so that the the numbers do not hug the edge of the input area. Also, removed toFixed() on operate() so whole number answers are not displayes with several zeroes behind a decimal point.